src/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
contextlib2==21.6.0
Django==5.1.7
django-ninja==1.4.3
django-ninja-extra==0.30.1
injector==0.22.0
pillow==11.1.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
django-cors-headers==4.6.0
django-ninja-jwt[crypto]
requests